feat: add memory-efficient transaction report using Java NIO streams

+++ zenon-fraud-detector/src/main/java/br/com/zenon/fraud/Transaction.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.Objects;
 
 public record Transaction(
-        Long step,
+        int step,
         TransactionType type,
         BigDecimal amount,
         TransactionCustomer origin,
